refactor(ml): extract mrm compliance helper to keep mrm_dossier under 500 LOC
The Codex-fix commits in this branch added ~120 LOC to mrm_dossier.py
(compliance status helper + thresholds + reason formatting), pushing it
from 391 to 511 LOC and tripping the ml_engine quality bar's global
500-LOC cap (tools/check_file_sizes.py).

Per the Arm C philosophy in
docs/superpowers/specs/2026-04-18-arm-c-ml-engine-quality-bar-design.md
("Each Arm C phase-PR deletes its target from this map"), splitting is
preferred over raising caps. Extracted _compliance_status,
PSI_FAIL_THRESHOLD, and ECE_FAIL_THRESHOLD into a new sibling module
apps.ml_engine.services.mrm_compliance — small (90 LOC), focused, and
pure-functional (no Django dependency).

Mechanical move only — no behavioural change. Tests updated to import
the helper from its new home; all 26 mrm tests pass. mrm_dossier.py is
now 437 LOC, comfortably under the cap.
